refactor(citizens): split per-target jobs tab into per-section files
Move the 962-LOC monolith into a folder:
- index.tsx: orchestrator (~150 LOC)
- AssignDialog.tsx: shared dialog
- Shared.tsx: CitizenTags, TargetRowShell, CollapsibleSection
- DepositsSection.tsx: DepositsSection + DepositTargetRow
- HusbandryCullingSection.tsx: HusbandrySection + CullingSection + PopulationTargetRow
- TradeRoutesSection.tsx: TradeRoutesSection + local/remote end rows

No behaviour change; all existing tests pass.

fix #419
